requirements.txt
pandas==2.2.2
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyYAML==6.0.1
schedule==1.2.1
